cong ctrl in progress
git-svn-id: https://ttuki.vtt.fi/svn/p2p-next/TUD/p2tp/trunk@499 e16421f0-f15b-0410-abcd-98678b794739

+++ ext/send_control.h
@@ -77,7 +77,7 @@ struct KeepAliveController : public SendController {
 
 struct CwndController : public SendController {
     
-    float   cwnd_;
+    double   cwnd_;
     
     CwndController(SendController* orig, int cwnd=1) :
     SendController(orig), cwnd_(cwnd) {    }
